DEV: restructure tests: move controllers
